backend/gallery-photo-download/requirements.txt
boto3==1.26.137
botocore==1.29.137
psycopg2-binary==2.9.9